Week2/story-generator/requirements.txt
streamlit==1.32.0
groq==0.4.2
python-dotenv==1.0.0